User access, change password

+++ src/audit-trail/audit-trail.controller.ts
@@ -1,11 +1,15 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, Query } from '@nestjs/common';
+import { Controller, Get, Post, Body, Patch, Param, Delete, Query, Request, HttpException, HttpStatus } from '@nestjs/common';
+import { UsersService } from 'src/users/users.service';
 import { AuditTrailService } from './audit-trail.service';
 import { CreateAuditTrailDto } from './dto/create-audit-trail.dto';
 import { UpdateAuditTrailDto } from './dto/update-audit-trail.dto';
+import { Role } from 'src/enums/role.enum';
 
 @Controller('audit-trail')
 export class AuditTrailController {
-  constructor(private readonly auditTrailService: AuditTrailService) {}
+  constructor(
+    private readonly auditTrailService: AuditTrailService,
+    private readonly usersService: UsersService) {}
 
   @Post()
   create(@Body() createAuditTrailDto: CreateAuditTrailDto) {
@@ -23,12 +27,22 @@ export class AuditTrailController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateAuditTrailDto: UpdateAuditTrailDto) {
-    return this.auditTrailService.update(+id, updateAuditTrailDto);
+  update(@Request() req: any, @Param('id') id: string, @Body() updateAuditTrailDto: UpdateAuditTrailDto) {
+    const user = req.user;
+    if(this.usersService.validateUserAccess(user, [Role.Admin, Role.Maintenance, Role.Supervisor])){
+      return this.auditTrailService.update(+id, updateAuditTrailDto);      
+    } else {
+      throw new HttpException("No tiene nivel suficiente", HttpStatus.FORBIDDEN);
+    }
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.auditTrailService.remove(+id);
+  remove(@Request() req: any, @Param('id') id: string) {
+    const user = req.user;
+    if(this.usersService.validateUserAccess(user, [Role.Admin, Role.Maintenance, Role.Supervisor])){
+      return this.auditTrailService.remove(+id);      
+    } else {
+      throw new HttpException("No tiene nivel suficiente", HttpStatus.FORBIDDEN);
+    }
   }
 }
